Jetson Nano Codes
These are the codes to be run by the jetson nano on Stocky. The light sensor code will allow you to threshold for red green blue and yellow lights and publish the detected color to ROS. The colors detected will tell the line tracking code which color of line to follow (red, green, blue) or to stop tracking lines and publishing command velocities (yellow). The line tracker code allows you to threshold the ranges of lines you want it to follow (red, green, blue) and publishes command velocities to the tank control code. the code "rft_hsvrgb.py" is the thresholding code used to find the ranges.